# backend/docs/swagger.yaml
basePath: /
definitions:
...
  service.AdminStudentDetail:
    properties:
      durationDays:
        type: integer
      fullName:
        type: string
      id:
        type: integer
      phone:
        type: string
      planTitle:
        type: string
      planType:
        description: '"workout" | "nutrition" | "both"'
        type: string
      remainingDays:
        type: integer
      restDays:
        items:
          type: string
        type: array
      startDate:
        type: string
      status:
        description: '"pending" | "active"'
        type: string
      weekly:
        items:
          type: string
        type: array
    type: object
  service.AdminStudentItem:
    properties:
      fullName:
        type: string
      id:
        type: integer
      phone:
        type: string
      planTitle:
        type: string
      planType:
        description: '"workout" | "nutrition" | "both"'
        type: string
      restDays:
        items:
          type: string
        type: array
      status:
        description: '"pending" | "active"'
        type: string
      weekly:
        items:
          type: string
        type: array
    type: object
  service.AdminStudentListResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/service.AdminStudentItem'
        type: array
      page:
        type: integer
      pageSize:
        type: integer
      total:
        type: integer
    type: object
  service.AdminStudentPatchRequest:
    properties:
      planId:
        type: integer
      status:
        description: '"pending" | "active"'
        type: string
    type: object
...
  service.DashboardStatsResponse:
    properties:
      activeUsers:
        type: integer
      purchasedCourses:
        type: integer
      totalUsers:
        type: integer
      year:
        type: integer
    type: object
  service.MonthlySaleItem:
    properties:
      courses:
        type: integer
      month:
        type: string
      sales:
        type: integer
    type: object
info:
  contact: {}
  description: Fitness Management System API documentation.
  title: Fitness CMS API
  version: "1.0"
paths:
  /admin/dashboard/monthly-sales:
    get:
      description: Returns 12 items (Persian month names) with courses count and sales
        sum for chart
      parameters:
      - description: Calendar year
        in: query
        name: year
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/service.MonthlySaleItem'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get monthly sales data for a year
      tags:
      - admin-dashboard
  /admin/dashboard/stats:
    get:
      description: Returns totalUsers, activeUsers, purchasedCourses for the given
        year (for admin dashboard)
      parameters:
      - description: Calendar year (e.g. 1403 or 2024)
        in: query
        name: year
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/service.DashboardStatsResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get dashboard stats for a year
      tags:
      - admin-dashboard
  /admin/students:
    get:
      description: Returns paginated list of students with status, plan info (for
        "شاگردهای من")
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: pageSize
        type: integer
      - description: 'Filter: all | pending | active'
        in: query
        name: status
        type: string
      - description: Search by name or phone
        in: query
        name: query
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/service.AdminStudentListResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List students for admin
      tags:
      - admin-students
  /admin/students/{id}:
    get:
      description: Returns one student with plan and optional startDate, durationDays,
        remainingDays
      parameters:
      - description: Student (user) ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/service.AdminStudentDetail'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get student detail for admin
      tags:
      - admin-students
    patch:
      consumes:
      - application/json
      description: 'PATCH body: status (pending|active) and/or planId to assign a
        plan'
      parameters:
      - description: Student (user) ID
        in: path
        name: id
        required: true
        type: integer
      - description: Optional status and planId
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/service.AdminStudentPatchRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 'message: ok'
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update student (status and/or assign plan)
      tags:
      - admin-students
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"